+++ lib/Payever/Core/Apm/Events/Context/Request/HeadersEntity.php
@@ -1,14 +1,12 @@
 <?php
 
 /**
- * PHP version 5.4 and 8.1
- *
  * @category  Apm Agent
  * @package   Payever\Core
  * @author    payever GmbH <service@payever.de>
  * @copyright 2017-2023 payever GmbH
  * @license   MIT <https://opensource.org/licenses/MIT>
- * @link      https://docs.payever.org/shopsystems/api/getting-started
+ * @link      https://docs.payever.org/api
  */
 
 namespace Payever\Sdk\Core\Apm\Events\Context\Request;