+++ lib/Payever/Core/Helper/DataHelper.php
@@ -1,14 +1,12 @@
 <?php
 
 /**
- * PHP version 5.4 and 8
- *
  * @category  Helper
  * @package   Payever\Core
  * @author    payever GmbH <service@payever.de>
- * @copyright 2017-2021 payever GmbH
+ * @copyright 2017-2026 payever GmbH
  * @license   MIT <https://opensource.org/licenses/MIT>
- * @link      https://docs.payever.org/shopsystems/api/getting-started
+ * @link      https://docs.payever.org/api
  */
 
 namespace Payever\Sdk\Core\Helper;
@@ -27,6 +25,7 @@ class DataHelper
      * @param mixed $data
      * @param string $className
      * @return RequestEntity|object|null
+     * @SuppressWarnings(PHPMD.StaticAccess)
      */
     public static function getEntityInstance($data, $className)
     {